moving disp.py closer to hillas.py
